fix: break circular dependency
